split orz.rs from mod
